requirements.txt
anyio==4.12.1
build==1.4.0
CacheControl==0.14.4
cachetools==7.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
dulwich==1.0.0
faiss-cpu==1.13.2
fastjsonschema==2.21.2
filelock==3.20.3
findpython==0.7.1
fsspec==2026.1.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.2
orjson==3.11.7
packaging==26.0
pandas==3.0.0
pbs-installer==2026.1.27
pkginfo==1.12.1.2
platformdirs==4.5.1
poetry==2.3.2
poetry-core==2.3.1
protobuf==6.33.5
pymilvus==2.6.8
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tomlkit==0.14.0
torch==2.10.0
trove-classifiers==2026.1.14.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.36.1
wheel==0.46.3
zstandard==0.25.0
